Add combinatorial borders between terrain types
Since the client's hex support is lame and we're ambitious, we need to generate
loads of sprites. The generated tileset is huge and uses a lot of memory, but
hopefully we can improve this in the future. Auto-generating sprites at least
allows us to simply change the generated layout without redrawing everything.