Final say on the corner generator
Generate all corner sprites as separate files to avoid making big files that
one cannot edit properly. Do it outside the vectron directory so we don't try
to render them twice. Switch to rsvg-convert because it's much faster than
Inkscape; for this to work, save the generator as "Simple SVG" instead of
"Inkscape SVG".